==> notebooks/3_evaluation.ipynb <==
# Predictions Evaluation

from pathlib import Path
import numpy as np
import pandas as pd

from utils.smooth_bleu import bleu_fromstr

def analyze_preds(base_file, sample_size=5):
    # read files
    hf_preds_file = Path(base_file).with_suffix('.hf_pred.csv')
    fine_tuned_file = Path(base_file).with_suffix('.finetuned_pred.csv')
    hf_preds = pd.read_csv(hf_preds_file)
    fine_tuned = pd.read_csv(fine_tuned_file)
    # put in df
    df = pd.DataFrame({'code': fine_tuned['code'],
     'hf_pred': hf_preds['prediction'],
     'fine_tuned_pred': fine_tuned['prediction'],
     'target': fine_tuned['target']})
    df.replace(np.nan, '', regex=True)
    # print sample with predictions
    sample = df.sample(sample_size, random_state=42)
    for code, hf_pred, fine_tuned_pred, target in sample.to_numpy():
        print('-------------------')
        print(code)
        print(f'## Human Review: {target}')
        print(f'## HF Pred: {hf_pred}')
        print(f'## Fine Tuned Pred: {fine_tuned_pred}')
    return df

def calc_bleu(df):
    refs = list(df['target'])
    preds = list(df['prediction'])
    for i in range(len(preds)):
        chars = "(_)`."
        for c in chars:
            preds[i] = preds[i].replace(c, " " + c + " ")
            preds[i] = " ".join(preds[i].split())
            refs[i] = refs[i].replace(c, " " + c + " ")
            refs[i] = " ".join(refs[i].split())
    return bleu_fromstr(preds, refs, rmstop=False)

def calc_bleu_score(base_file):
    hf_preds_file = Path(base_file).with_suffix('.hf_pred.csv')
    fine_tuned_file = Path(base_file).with_suffix('.finetuned_pred.csv')
    hf_preds = pd.read_csv(hf_preds_file)
    ft_preds = pd.read_csv(fine_tuned_file)
    hf_preds.replace(np.nan, '', regex=True, inplace=True)
    ft_preds.replace(np.nan, '', regex=True, inplace=True)
    hf_bleu = calc_bleu(hf_preds)
    ft_bleu = calc_bleu(ft_preds)
    print(f'HF BLEU: {hf_bleu}')
    print(f'Fine Tuned BLEU: {ft_bleu}')
    return hf_bleu, ft_bleu

## Qualitative Evaluation
We will now compare the predictions of the HF model and the fine-tuned model on samples of the four datasets.

We will print the code, the prediction of the HF model and the prediction of the fine-tuned model.

df = {}

df['msg'] = analyze_preds('../data/msg-test')

df['vscode'] = analyze_preds('../data/microsoft_vscode_1000.csv')

df['kotlin'] = analyze_preds('../data/JetBrains_kotlin_1000.csv')

df['uppy'] = analyze_preds('../data/transloadit_uppy_1000.csv')

As we can see, the models are pretty mediocre at understanding the issue. But I find that the HF model tends to produce more generic predictions, while the fine-tuned model produces predictions that are more specific to the code and shows a better understanding.

Still, both models predict something sensible but struggle to pin-point the problem.

## Quantitative Evaluation
For each dataset, we calculate the [BLEU-4](https://en.wikipedia.org/wiki/BLEU) score for the predictions of the HF model and the fine-tuned model. The BLEU score is a measure of how similar the predictions are to the target. The higher the score, the better the predictions.

calc_bleu_score('../data/msg-test')

calc_bleu_score('../data/microsoft_vscode_1000.csv')

calc_bleu_score('../data/JetBrains_kotlin_1000.csv')

calc_bleu_score('../data/transloadit_uppy_1000.csv')

As we can see, the fine-tuned model performs slightly better than the HF model on all datasets.

Nevertheless, the score is still pretty low (as the authors of {cite}`li2022codereviewer` put it: "it is a hard task").
